change error classes directory
