+++ core/src/core/error.ts
@@ -1,59 +1,7 @@
 import store from '../../store';
-import { ErrorType } from '../../types/error';
+import { ErrorType, GeneralError } from '../../types/error';
 import { MiddleWare, MiddleWareError } from '../../types/global';
 import logger from '../handlers/log.handler';
-
-export class GeneralError extends Error {
-  isGeneral = true;
-  type: string = ErrorType.General;
-  constructor(
-    public message: string,
-    public code: number,
-    type?: string,
-    public additionType?: string
-  ) {
-    super(message);
-    this.type = type ?? ErrorType.General;
-    if (additionType) this.type = `${this.type}:${additionType}`;
-  }
-}
-
-export class BadRequestError extends GeneralError {
-  constructor(message: string, additionType?: string) {
-    super(message, 400, ErrorType.BadRequest, additionType);
-  }
-}
-export class ForbiddenError extends GeneralError {
-  constructor(message: string, additionType?: string) {
-    super(message, 403, ErrorType.Forbidden, additionType);
-  }
-}
-export class UnauthorizedError extends GeneralError {
-  constructor(message: string, additionType?: string) {
-    super(message, 401, ErrorType.Unauthorized, additionType);
-  }
-}
-
-export class DuplicateError extends BadRequestError {
-  isDuplicate = true;
-  constructor(message: string) {
-    super(message, ErrorType.Duplication);
-  }
-}
-
-export class ValidationError extends BadRequestError {
-  isValidation = true;
-  constructor(message: string, additionalType?: ErrorType) {
-    super(message, additionalType);
-  }
-}
-
-export class LimitError extends GeneralError {
-  isLimit = true;
-  constructor(message: string) {
-    super(message, 429, ErrorType.Limitation);
-  }
-}
 
 export const errorHandler: MiddleWareError = (error, req, res, next) => {
   let code = 500,
